Wrapping terms of the form Vec<T> into Wrapper(T) in order to be able to have a fine-grained control about how it is encoded/read.
Vec<T> is encoded as the series of the encoding of the items (without length prefix). Wrapper(T) is encoded by adding a length prefix whose teh exact format depends on T/Wrapper. For bit-level mutations, this also improves things: (i) mutate Wrapper can mutate the length of the list, mutate Vec<T> can alter the structure of the encoded vector and (ii) we can more easily find where to replace payloads when scanning sub-terms of the form Vec<T>.

- impacted/new types: Compressions, CipherSuites, ClientExtensions, ServerExtensions, HelloRetryExtensions, CertificateEntries.
- rustls::msgs:message.rs now uses those new encode/read implems for those types
- adding to the signature symbols/functions like fn_cipher_suites_make that tranform Vec<T> into Wrapper(T)
- modify seeds and vulns accordingly
- tests pass, including the new `test_replace_bitstring_multiple`