PUT rearchitecture: build output cleanup (#297)
* build: remove metadata duplication between packages
* style: apply consistent style to C code
* build: fix rustdoc warnings on nyi_fn macro
* build: fix unknown lint directive `clippy`
* build: replace deprecated calls to bindgen's rustfmt_bindings
* build: remove warnings related to exposing ffi
* refactor: split bindings code of openssl into version-specific modules
* refactor: fix warnings in puffin and tlspuffin packages
* refactor: fix warnings in wolfssl-related code
* refactor: fix warnings in sshpuffin package
* refactor: fix warnings about missing Default trait
* refactor(puffin): fix leftover dead code from PRs #283 and #284

* refactor: remove unnecessary supertrait Codec of ProtocolMessage
At this point, there is no reason to encode/read a ProtocolMessage. The actual on-the-wire message is always the opaque counterpart, itself providing its own Codec implementation. In every instance where the message is encoded/read, it always uses the opaque message as intermediary before trying to create a ProtocolMessage from that. We also remove the (unused) Codec implementation on Message.