build: avoid building vendor bindings per-default
We exclude the vendored crates from the workspace so that they are not built by
default when running cargo without specifying a package. They will only be
pulled in when requested by a workspace member (tlspuffin, sshpuffin, ...).

This should marginally reduce build time during development and make sure that
these crates are correcly configured by the requesting workspace member when
pulled in the build, since the default features don't usually match our need.