build: fail early when both openssl and wolfssl bindings are selected
When building tlspuffin with both openssl and wolfssl selected through feature
flags, the build process progress a lot before finally reaching a compile-time
error in the puffin package.

This set the compile-time check in the build script of the tlspuffin package,
so that the error message happens in the package where these features are
defined.