build: use a consistent naming scheme for vendor versions
In vendored crates, the feature flags for versions are sometime prefixed with
`vendored-` (libssh, wolfssl, libressl) and sometime not (openssl), making them
inconsistent with the user-facing features defined in the tlspuffin/sshpuffin
package. This prefix is also redundant in our fuzzing use case: we always want
to fuzz specific versions of these libraries and never the system ones.

We remove the `vendored-` prefix from all the features flags, to keep the
naming scheme consistent.