fix build when incompatible features are requested (#303)
* build: avoid building vendor bindings per-default

We exclude the vendored crates from the workspace so that they are not built by
default when running cargo without specifying a package. They will only be
pulled in when requested by a workspace member (tlspuffin, sshpuffin, ...).

This should marginally reduce build time during development and make sure that
these crates are correcly configured by the requesting workspace member when
pulled in the build, since the default features don't usually match our need.

* build: fail when incompatible openssl (resp. wolfssl) features are requested

Currently, when several versions of openssl or wolfssl are requested through
cargo, the vendored crates will silently choose one version. Additionally, the
PUT capabilities (session resumption, tls3 support, ...) associated with ALL
these features will be activated in tlspuffin at the same time and trigger hard
to decypher build errors, or worse, they might actually build correctly if most
features are compatible but result in unexpected behavior at runtime.

This fixes the problem by testing that one and only one version of openssl
(resp. wolfssl) sources can be pulled in the build through the use of feature
flags.

to squash: reject unknown openssl/wolfssl/libressl/libssh versions

* build: fail early when both openssl and wolfssl bindings are selected

When building tlspuffin with both openssl and wolfssl selected through feature
flags, the build process progress a lot before finally reaching a compile-time
error in the puffin package.

This set the compile-time check in the build script of the tlspuffin package,
so that the error message happens in the package where these features are
defined.

* build: use a consistent naming scheme for vendor versions

In vendored crates, the feature flags for versions are sometime prefixed with
`vendored-` (libssh, wolfssl, libressl) and sometime not (openssl), making them
inconsistent with the user-facing features defined in the tlspuffin/sshpuffin
package. This prefix is also redundant in our fuzzing use case: we always want
to fuzz specific versions of these libraries and never the system ones.

We remove the `vendored-` prefix from all the features flags, to keep the
naming scheme consistent.

* build: remove references to submodules (removed in #243)