feat(openssl): implement non-clear mode for OpenSSL PUT
During PUT OpenSSL initialization a confusing message about the
non-availability of clear mode is logged, but only when clear mode is NOT
selected. In fact, the OpenSSL PUT only implements clear mode rather than a
full reset of the agents.

This is confusing on two levels:
- the message hints that a full reset of OpenSSL agents is always performed but
  they are, on the contrary, always soft reset.
- the process continues with the soft reset without making it clear to the user
  that the option they have selected won't be applied.

We implement the non-clear mode in the OpenSSL PUT to get rid of the message.

Fixes: #154