fix(tlspuffin): don't generate incompatible seeds when there is no built-in PUT
When no PUT is selected at build-time, the `tlspuffin seed` command still
generates seeds. These seeds are incompatible with the default TCP PUT and
`tlspuffin execute` fails on this initial corpus.

We fix the `tlspuffin seed` command to make sure that it only generates seeds
that are compatible with the build-time configuration of tlspuffin.