fix(cli): correct default `-n <num>` value when running `execute` on directories
When the `execute` subcommand is run, it executes by default all input traces
provided on the command line. This is done by defaulting the `-n <num>` CLI
argument to the number of provided arguments.

When a directory is part of the command line inputs, all input traces in that
directory will be executed. Since the number of inputs doesn't match the number
of arguments on the command line anymore, `execute` will wrongly ignore part of
the inputs and only run on a subset of the provided inputs.

We fix this inconsistency by using the total number of traces found as default
value for `-n <num>`.