refactor(puffin): remove 'static lifetime constraint on PUT registry
The fuzzer and the tests rely on the presence of a global PUT registry which is
used to instantiate agents during a trace execution. Apart from creating a
shared global state between modules, this is impractical to register a PUT
dynamically and thus to support PUTs written in C that are only available at
link-time.

We remove the constraint and pass where necessary a default registry whose
content matches the previously statically allocated registry.