QoL changes for better CI integration (#304)
* feat(openssl): implement non-clear mode for OpenSSL PUT

During PUT OpenSSL initialization a confusing message about the
non-availability of clear mode is logged, but only when clear mode is NOT
selected. In fact, the OpenSSL PUT only implements clear mode rather than a
full reset of the agents.

This is confusing on two levels:
- the message hints that a full reset of OpenSSL agents is always performed but
  they are, on the contrary, always soft reset.
- the process continues with the soft reset without making it clear to the user
  that the option they have selected won't be applied.

We implement the non-clear mode in the OpenSSL PUT to get rid of the message.

Fixes: #154

* fix(cli): don't use tlspuffin exit code to convey semantically meaningful data

The `tlspuffin execute` command exits with 0 or 1 to signal if there are
potentially more inputs that were not executed. This has been used for batch
executions during coverage analysis.

Numerous tools, including the running shell, associate a single meaning to
non-zero exit code: the command failed to execute in the way intended by the
caller. Depending on how they are configured, these tools might behave
differently when they detect a non-zero return code (e.g. bash "set -e").

To avoid any confusion, we instead log the total number of inputs, so that
calling scripts can still retrieve it without relying on the exit code.

* fix(cli): display missing trace file when the `execute` command fails

* docs(cli): fix seeds directory name in CLI help and documentation

* fix(tlspuffin): don't generate incompatible seeds when there is no built-in PUT

When no PUT is selected at build-time, the `tlspuffin seed` command still
generates seeds. These seeds are incompatible with the default TCP PUT and
`tlspuffin execute` fails on this initial corpus.

We fix the `tlspuffin seed` command to make sure that it only generates seeds
that are compatible with the build-time configuration of tlspuffin.

* fix(cli): remove duplicated log message when generating seeds

* fix(cli): correct default `-n <num>` value when running `execute` on directories

When the `execute` subcommand is run, it executes by default all input traces
provided on the command line. This is done by defaulting the `-n <num>` CLI
argument to the number of provided arguments.

When a directory is part of the command line inputs, all input traces in that
directory will be executed. Since the number of inputs doesn't match the number
of arguments on the command line anymore, `execute` will wrongly ignore part of
the inputs and only run on a subset of the provided inputs.

We fix this inconsistency by using the total number of traces found as default
value for `-n <num>`.

* refactor(openssl): fix clippy errors in openssl-src

* revert(cli): revert changes to `execute` command, introduce `execute-traces` instead

We introduce a new command to avoid breaking coverage workflows. In the future,
we want to merge these commands but this will require having stable coverage
scripts first.