fix(openssl): correctly reseed OpenSSL's deterministic RNG
When we reimplemented the deterministic RNG for OpenSSL, we enforced a minimum
number of 8 bytes for the RNG seed. Because the change was not propagated to
the Rust wrapper code, we kept trying to reseed the RNG with a 4-bytes buffer,
which had no effect.

This commit fixes the default seed passed from Rust in order to correctly
reseed the deterministic RNG. It also modifies the corresponding unit test to
prevent future regressions of this problem.