Fix find_unique_match heuristic 3
find_relative_node now returns `shift_ancestor_to_search:usize`: this can be non-zero when p is not a sibling but a sibling of an ancestor of to_search.
It then corresponds to the position of `to_search` relatively to the evaluation of this ancestor.
This can happen for example for append(f, fn_support_group_extension(to_search)) when relative node is f and fn_support_group_extension add some headers in front of to_search. shift_ancestor_to_search will be the length of this header.