build: fix harness build failing depending on the current locale
The harness build process currently relies on a `glob` search to gather all the
static libraries that are part of the vendor library. Depending on the locale
set, the order of the libraries returned by this search varies. This in turn
leads to libraries being linked in varying orders potentially resulting in
missing symbols in the final PUT (e.g. libssl depends on symbols in libcrypto
and should therefore be linked first).

To avoid this problem, this commit keeps all the symbols in the linked
archives, so that the link order of these libraries becomes irrelevant.