fix: preserve X509 stores when crossing ffi boundary
X509 stores sent to the C PUTs agents were corrupted and the memory locations
pointing to the certificates become invalid, crashing the fuzzer. This only
happens in release builds e.g.:

```
cargo test --release -p tlspuffin --features=cputs,openssl312
```

This seems to be due to the Rust compiler (wrongly?) optimizing away the array
slices passed to `make_descriptor`. Surprisingly, simply iterating over these
slices in `make_descriptor` solves the problem.

This commit keep certificates in a local variable array instead of creating a
slice in-place, forcing the Rust compiler to preserve the slice during all the
duration of the call to `make_descriptor`.